Map input data directly into memory
The purpose of reading the input data from a stream into
a buffer is to avoid I/O noise during the measurement proper.
However, this method does not guarantee intervention by the
kernel, and can be done more reliably by mapping the input
data (on Linux) with MAP_POPULATE. If allocations are properly
lockeds, this guarantees that the data will remain in RAM
until explicit deallocation.

On systems that don't support MAP_POPULATE, we try to
place data in RAM after mapping on a best-effort basis
by touching every involved page frame once before
commencing with measurements.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Reviewed-by: Stefanie Scherzinger <stefanie.scherzinger@uni-passau.de>