Implement latency measurement boundaries
When measuring over extended periods of time, it is neither
useful nor possible (memory consumption) to log every single
latency value. Since usually only outliers are of interest,
allow for specifying an upper and lower latency bound
on the command line above/below which latencies are recorded,
and discard all "normal" values in between.

Also add a command line option buffer-frac to limit the size
of the log buffer (which is otherwise allocated with a capabity
for the worst possible case of measured latency values)
to a certain fraction of the default size.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@othr.de>
Reviewed-by: Stefanie Scherzinger <stefanie.scherzinger@uni-passau.de>