Add ftrace low-level tracing support
When DBToaster experiences high latencies, this can either be
caused by the query code itself, or by (asynchronous or synchronous)
effects of the global system. A proper understanding of the latter
can only be gained if it is known what is going on in non-DBToaster
parts of the system. Introduce a command line switch --ftrace to
activate the ftrace tracer during query evaluation.
To ignore initial startup latencies, introduce --ftrace-rep to
set the amount of high latency threshold violations after which
a trace is stopped.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Reviewed-by: Stefanie Scherzinger <stefanie.scherzinger@uni-passau.de>